+++ cms/src/features/posts/Posts.tsx
@@ -5,19 +5,42 @@ import useAxiosPrivate from "../../hooks/useAxiosPrivate";
 import { PostSuccessResponse } from "../../interfaces/responses";
 import { useEffect, useState } from "react";
 import { PostFormError } from "../../interfaces/errors";
-import { useParams } from "react-router";
+import { useLoaderData, useParams } from "react-router";
+import { useAuth } from "../../hooks/useAuth";
 
 // Fetches authenticated user's posts
 //  and renders their posts
 // Should this fetch user based on :username parameters?
 const Posts = () => {
+  const [posts, setPosts] = useState([]);
+  const axiosPrivate = useAxiosPrivate();
   const params = useParams<{ username: string }>();
-
-  const { user, addPost } = useUserData();
+  console.group("Posts component rendering...");
+  console.log(posts);
+  /* const { user, addPost } = useUserData();
   const [errors, setErrors] = useState<PostFormError>();
   const data = user?.blog?.posts || []; // This should fetch based on :username parameter
-  const axiosPrivate = useAxiosPrivate();
-  /*const submitPost = async (data: any) => {
+  const axiosPrivate = useAxiosPrivate(); */
+
+  useEffect(() => {
+    const getUserPosts = async () => {
+      try {
+        const response = await axiosPrivate.get(
+          `users/${params.username}/posts`
+        );
+        setPosts(response.data.posts);
+      } catch (err) {
+        console.error(err);
+      }
+    };
+
+    getUserPosts();
+    console.group("Posts mounted...");
+    console.log(params);
+    console.groupEnd();
+  }, []);
+
+  /* const submitPost = async (data: any) => {
     try {
       console.group("submitPost running...");
       const response = await axiosPrivate.post<PostSuccessResponse>(
@@ -28,9 +51,9 @@ const Posts = () => {
     } catch (err: any) {
       setErrors(err.response.data.errors);
     }
-  };
+  }; */
 
-  return (
+  /* return (
     <>
       <h2>Posts</h2>
       <p>username {user?.username}</p>
@@ -38,14 +61,13 @@ const Posts = () => {
       <PostList data={data} />
     </>
   ); */
-
-  useEffect(() => {
-    console.group("Posts mounted...");
-    console.log(params);
-    console.groupEnd();
-  });
-
-  return <p>Posts</p>;
+  return (
+    <>
+      <h2>Posts</h2>
+      <p>username {params.username}</p>
+      <PostList data={posts} />
+    </>
+  );
 };
 
 export default Posts;